requirements.txt
pymupdf>=1.23.22
faiss-cpu>=1.8.0
langchain-text-splitters>=0.0.1
python-dotenv>=1.0.1
numpy>=1.26.4
reportlab>=4.4.10
tiktoken>=0.6.0
transformers>=4.38.1
torch>=2.2.1
scikit-learn>=1.4.1
datasets>=2.17.1
pytest>=8.0.2
openai>=1.12.0
accelerate>=0.26.0
